corrected type assignment

+++ src/oersted/testing.py
@@ -87,7 +87,7 @@ def make_helmholtz(size, jmag: None | float = None, scale=1e-3) -> tuple[Mesh, N
     nodes_lower[:, 2] -= 0.2  # flip to lower side
     nodes = np.vstack((nodes_upper, nodes_lower))
     connectivity_upper: NDArray[uint32] = ring_mesh.connectivity.copy()
-    connectivity_lower: NDArray[uint32] = ring_mesh.connectivity.copy() + ring_mesh.num_nodes
+    connectivity_lower: NDArray[uint32] = ring_mesh.connectivity.copy() + uint32(ring_mesh.num_nodes)
 
     helmholtz_mesh = Mesh(nodes, np.vstack((connectivity_upper, connectivity_lower)))
